refactor: move unused components to foyer-util
