+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "foyer-intrusive",
     "foyer-memory",
     "foyer-storage",
+    "foyer-util",
 ]
 
 [workspace.dependencies]
