+++ foyer-util/Cargo.toml
@@ -1,16 +1,18 @@
 [package]
-name = "foyer-common"
-version = "0.6.4"
+name = "foyer-utils"
+version = "0.0.0"
 edition = "2021"
 authors = ["MrCroxx <mrcroxx@outlook.com>"]
-description = "common components for foyer - the hybrid cache for Rust"
+description = "utils for foyer - the hybrid cache for Rust"
 license = "Apache-2.0"
 repository = "https://github.com/mrcroxx/foyer"
 homepage = "https://github.com/mrcroxx/foyer"
 readme = "../README.md"
+publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bitmaps = "3"
 bytes = "1"
 cfg-if = "1"
 crossbeam = "0.8"
