+++ foyer-util/src/batch.rs
