+++ foyer-util/src/lib.rs
@@ -14,17 +14,9 @@
 
 #![cfg_attr(coverage_nightly, feature(coverage_attribute))]
 
-pub mod async_batch_pipeline;
-pub mod asyncify;
-pub mod bits;
-pub mod buf;
-pub mod code;
-pub mod countdown;
-pub mod object_pool;
-pub mod range;
-pub mod rate;
-pub mod rated_ticket;
-pub mod runtime;
-
-#[cfg(any(target_os = "linux", target_os = "macos"))]
-pub mod fs;
+pub mod batch;
+pub mod continuum;
+pub mod erwlock;
+pub mod iostat;
+pub mod judge;
+pub mod slab;
